webrecruiter/requirements.txt
appnope==0.1.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
boto==2.40.0
boto3==1.3.1
botocore==1.4.18
decorator==4.0.9
defusedxml==0.4.1
Django==2.0.6
django-allauth==0.25.2
django-autofixture==0.12.1
django-bootstrap3==7.0.1
django-cart==1.0.4
django-countries==3.4.1
django-discover-runner==1.0
django-extensions==1.6.1
django-localflavor==1.2
django-phonenumber-field==1.1.0
django-role-permissions==2.2.0
django-simple-history==2.5.1
django-storages==1.4.1
django-wysiwyg-redactor==0.5.1
djangorestframework==3.9.0
docutils==0.12
gnureadline==6.3.8
html5lib==1.0.1
httplib2==0.9.2
ipython==4.2.0
ipython-genutils==0.1.0
jmespath==0.9.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pexpect==4.0.1
phonenumberslite==7.3.1
pickleshare==0.7.2
Pillow==5.2.0
ptyprocess==0.5.1
PyPDF2==1.26.0
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2018.4
reportlab==3.5.6
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
stripe==2.10.0
traitlets==4.2.1
webencodings==0.5.1
xhtml2pdf==0.2.2
